+++ flask_lastuser/sqlalchemy.py
@@ -8,18 +8,34 @@ SQLAlchemy extensions for Flask-Lastuser.
 
 from __future__ import absolute_import
 
-__all__ = ['UserBase', 'TeamBase', 'ProfileMixin', 'UserManager']
-
 import urlparse
 from pytz import timezone
 from werkzeug import cached_property
 from flask import g, current_app
-from sqlalchemy import Column, Boolean, Integer, String, Unicode, ForeignKey, Table, UniqueConstraint
+from sqlalchemy import (Column, Boolean, Integer, String, Unicode, ForeignKey, Table, UniqueConstraint,
+    MetaData)
 from sqlalchemy.orm import deferred, undefer, relationship, synonym
 from sqlalchemy.ext.declarative import declared_attr
 from flask.ext.lastuser import UserInfo, UserManagerBase
-from coaster import getbool, make_name
+from coaster.utils import getbool, make_name, LabeledEnum
 from coaster.sqlalchemy import BaseMixin, JsonDict, BaseNameMixin
+
+
+__all__ = ['UserBase', 'UserMergeMixin', 'TeamBase', 'ProfileMixin', 'UserManager', 'IncompleteUserMigration']
+
+
+class IncompleteUserMigration(Exception):
+    """
+    Could not migrate users because of data conflicts
+    """
+    pass
+
+
+class USER_STATUS(LabeledEnum):
+    ACTIVE = (0, 'Active')        # Currently active
+    SUSPENDED = (1, 'Suspended')  # Suspended upstream
+    MERGED = (2, 'Merged')        # Merged locally (all data migrated)
+    DELETED = (3, 'Deleted')      # Deleted but record preserved for data
 
 
 class UserBase(BaseMixin):
@@ -168,6 +184,129 @@ class UserBase(BaseMixin):
     user_organization_owned_ids = user_organizations_owned_ids
 
 
+class UserMergeMixin(object):
+    @declared_attr
+    def status(cls):
+        return Column(Integer, nullable=False, default=USER_STATUS.ACTIVE)
+
+    @property
+    def is_active(self):
+        """
+        Is the user active? This is local status, not upstream status from Lastuser.
+        """
+        return self.status == USER_STATUS.ACTIVE
+
+    @property
+    def is_suspended(self):
+        """
+        Is the user suspended? This is local status, not upstream status from Lastuser.
+        """
+        return self.status == USER_STATUS.SUSPENDED
+
+    @property
+    def is_merged(self):
+        """
+        Is the user merged? This is local status, not upstream status from Lastuser.
+        """
+        return self.status == USER_STATUS.MERGED
+
+    def merge_into(self, user):
+        """
+        Merge self into the specified user and relink all 
+        """
+        if self.status == USER_STATUS.MERGED:
+            return  # We are already merged, so ignore this call
+
+        assert isinstance(user, UserMergeMixin) and user != self
+        # TODO: Do stuff here
+
+        # User id column (for foreign keys)
+        user_id_column = self.__class__.__table__.c.id  # 'id' is from IdMixin via BaseMixin via UserBase
+        # Session (for queries)
+        session = self.query.session
+
+        # Keep track of all migrated tables
+        migrated_tables = set()
+        safe_to_remove_user = True
+
+        # Find the Base class
+        base = self.__class__
+        while True:
+            goparent = False
+            for cbase in base.__bases__:
+                if hasattr(cbase, 'metadata') and isinstance(cbase.metadata, MetaData):
+                    base = cbase
+                    goparent = True
+                    break
+            if not goparent:
+                break
+
+        def do_migrate_table(table):
+            user_columns = []
+            for column in table.columns:
+                for fkey in column.foreign_keys:
+                    if fkey.column is user_id_column:
+                        # This table needs migration on this column
+                        user_columns.append(column)
+                        break
+                # Check for unique constraint on this column (single or multi-index)
+                # If so, return False
+                if column.unique:
+                    # XXX: This will fail for secondary relationship tables, which will have a unique index
+                    # but no model on which to place migrate_user, unless one of the related models handles
+                    # migrations AND signals a way for this table to be skipped here. This is why 
+                    # model.migrate_user below returns a list of table names it has processed.
+                    return False
+
+            for constraint in table.constraints:
+                if isinstance(constraint, UniqueConstraint):
+                    for column in constraint.columns:
+                        if column in user_columns:
+                            # user_id is part of a unique constraint. We can't migrate automatically.
+                            return False
+
+            # TODO: If this table uses Flask-SQLAlchemy's bind_key mechanism, session.execute won't bind
+            # to the correct engine, so the table cannot be migrated. If we attempt to retrieve and connect
+            # to the correct engine, we may lose the transaction. We need to confirm this.
+            if table.info.get('bind_key'):
+                return False
+
+            for column in user_columns:
+                session.execute(table.update().where(column==self.id).values(**{column.name:user.id}))
+                session.flush()
+            return True
+
+        # Look up all subclasses of this base class
+        for model in base.__subclasses__():
+            if model != self.__class__:
+                if hasattr(model, 'migrate_user'):
+                    try:
+                        result = model.migrate_user(olduser=self, newuser=user)
+                        if isinstance(result, (list, tuple, set)):
+                            migrated_tables.update(result)
+                        migrated_tables.add(model.__table__.name)
+                    except IncompleteUserMigration:
+                        safe_to_remove_user = False
+                else:
+                    # No model-backed migration. Figure out all foreign key references to user table
+                    if not do_migrate_table(model.__table__):
+                        safe_to_remove_user = False
+                    migrated_tables.add(model.__table__.name)
+
+        # Now look in the metadata for any tables we missed
+        for table in base.metadata.tables.values():
+            if table.name not in migrated_tables:
+                if not do_migrate_table(table):
+                    safe_to_remove_user = False
+                migrated_tables.add(table.name)
+
+        # Finally, release claim to username and email (unique properties) and mark self as merged
+        self.username = None
+        self.email = None
+        if safe_to_remove_user:
+            self.status = USER_STATUS.MERGED
+
+
 class TeamBase(BaseMixin):
     __tablename__ = 'team'
 
@@ -190,6 +329,33 @@ class TeamBase(BaseMixin):
     @declared_attr
     def users(cls):
         return relationship('User', secondary='users_teams', backref='teams')
+
+    @classmethod
+    def migrate_user(cls, olduser, newuser):
+        """
+        Substitute the old user's team membership with the new user and return the list of
+        tables affected.
+        """
+        session = cls.query.session
+        users_teams = cls.metadata.tables['users_teams']
+
+        affected_team_ids = set([r.team_id for r in session.query(
+            users_teams.c.team_id).filter_by(user_id=olduser.id).all()])
+
+        # newuser is already in these teams
+        unaffected_team_ids = set([r.team_id for r in session.query(
+            users_teams.c.team_id).filter_by(user_id=newuser.id).all()])
+
+        migrate_team_ids = affected_team_ids - unaffected_team_ids
+        remove_team_ids = affected_team_ids.intersection(unaffected_team_ids)
+
+        session.execute(users_teams.update().where(
+            users_teams.c.user_id == olduser.id, users_teams.c.team_id.in_(migrate_team_ids)).values(user_id=newuser.id))
+        session.execute(users_teams.delete(
+            users_teams.c.user_id == olduser.id, users_teams.c.team_id.in_(remove_team_ids)))
+
+        # We handled migrations in the users_teams table, so let the caller know
+        return ['users_teams']
 
 
 class ProfileMixin(object):
@@ -295,6 +461,46 @@ class ProfileMixin(object):
                         setattr(profile, type_col, type_org)
                     session.add(profile)
 
+        # Fourth, migrate profiles if there are any matching the user's old ids
+        if user.oldids:
+            profile = cls.query.filter_by(userid=user.userid).first()
+            if profile:
+                oldprofiles = cls.query.filter(cls.userid.in_(user.oldids)).all()
+                for op in oldprofiles:
+                    op.migrate_to(profile)
+
+    def update_from_lastuser(self):
+        """
+        Query Lastuser for current details of this userid and update as necessary.
+        """
+        lastuser = current_app.extensions.get('lastuser')
+        if lastuser:
+            userinfo = lastuser.getuser_by_userid(self.userid)
+            if userinfo:
+                if userinfo['userid'] != self.userid and self.userid in userinfo.get('oldids', []):
+                    # This Profile has gone away. Does the new profile exist here?
+                    profile = self.query.filter_by(userid=userinfo['userid']).first()
+                    if profile:
+                        self.migrate_to(profile)
+                    else:
+                        # The new profile isn't here yet, so assume their identity
+                        self.userid = userinfo['userid']
+                moveprofile = self.query.filter_by(name=userinfo['name']).first()
+                if moveprofile and moveprofile != self:
+                    # There's another profile holding our desired name. Move it out of the way
+                    moveprofile.name = moveprofile.userid
+                self.name = userinfo['name']
+                self.title = userinfo['title']
+
+
+    def migrate_to(self, profile):
+        """
+        Move all data from self to the other profile, typically when merging user accounts
+        """
+        pass
+        # TODO
+        # Uncomment after implementation of details:
+        # session.delete(self)
 
 
 class ProfileBase(ProfileMixin, BaseNameMixin):
@@ -302,73 +508,7 @@ class ProfileBase(ProfileMixin, BaseNameMixin):
     Base class for profiles
     """
     userid = Column(Unicode(22), nullable=False, unique=True)
-
-
-class UserMigrateMixin(object):
-    """
-    UserMigrateMixin provides helper methods to handle user data migration when user
-    accounts are merged. It depends on the class having a ``user_id`` column that points
-    to the ``user`` table.
-    """
-    @classmethod
-    def _get_user_id_unique_with(cls):
-        """
-        Return the user_id column and the other columns it's unique with
-        """
-        if 'user_id' not in cls.__table__.c:  # pragma: no cover
-            return None, []  # This table does have a user_id column
-        user_id_col = cls.__table__.c.user_id
-        unique_with = []
-        if not user_id_col.primary_key and not user_id_col.unique:
-            # user_id is present but isn't a primary key or unique by itself.
-            # Is there a unique constraint involving user_id? Find the other columns
-            for constraint in cls.__table__.constraints:
-                if isinstance(constraint, UniqueConstraint):
-                    candidate = False
-                    other_columns = []
-                    for column in constraint.columns:
-                        if column == user_id_col:
-                            candidate = True
-                        else:
-                            other_columns.append(column)
-                    if candidate:
-                        unique_with.extend(other_columns)
-        return user_id_col, unique_with
-
-    @classmethod
-    def migrate_user_conflicts(cls, olduser, newuser):
-        """
-        Return rows with conflicting data when migrating from olduser to newuser.
-        This involves checking for unique constraints on the ``user_id`` column.
-
-        If this model has no ``user_id`` or no unique constraint on ``user_id``,
-        an empty list is returned.
-
-        :returns: List of 2-tuples of conflicting rows
-        """
-        user_id_col, unique_with = cls._get_user_id_unique_with()
-        if user_id_col is None:
-            return []
-        # TODO
-
-    @classmethod
-    def migrate_user(cls, olduser, newuser, discard=[]):
-        """
-        Merge data for olduser into newuser. ``discard`` should be a list of row ids
-        to be discarded.
-        """
-        user_id_col, unique_with = cls._get_user_id_unique_with()
-        if user_id_col is None:
-            return
-        for row in discard:
-            cls.query.filter_by(id=row).delete()
-        # TODO
-
-    def merge_data_from(self, other):
-        """
-        Merge data from the other instance.
-        """
-        raise NotImplementedError("Subclasses must provide this method.")
+    status = Column(Integer, nullable=False, default=USER_STATUS.ACTIVE)
 
 
 def make_user_team_table(base):