+++ CHANGES.rst
@@ -4,6 +4,8 @@
 * Set Cache-Control header to 'private' to prevent proxy caching of
   logged-in pages.
 * requires_permission and has_permission now accept multiple permissions.
+* Userinfo is now stored as a JSON column where available.
+* User.phone now returns the primary verified phone number, if available.
 
 0.3.13
 ------